Do the big refactor for more elaboration tolerance
In particular, now it is much easier to:
* Share information between feature settings (gold QA-SRL, propbank, etc.)
* Specify new modular clustering model constructors
* Share code between data splits (train/dev/test)
* Add new feature data, including computed data cached on disk

A bunch of the old stuff (like evaluation and coindexing code) is commented out since it no longer
works. These things are mostly obsolete anyway, but may be useful as a reference when implementing
the new eval.

The frame browser has not been tested yet.