Add a more efficient agglomerative clustering algorithm
This takes N^3 in the worst case and N^2 in the best case, doing better
when the clusters are better separated by the distance function.
Basically, it's A* agglomerative clustering with lazy node expansion and
pruning, taking advantage of a particular monotonicity property of the
loss functions I'm using.

As it stands it does clustering of max 50 instances for all verbs in the
sanity-check setting in 9min as opposed to 12-13min for the previous
method. Hopefully adding constraints to question clustering will make
this faster as well, otherwise it's still way too slow..