WIP: OntoNotes fixes and Features refactor
I need to change the clustering away from using QuestionId, since
PropBank arguments are clustered by span, and not by question. This
may also make sense for QA-SRL clustering in e.g., the predicted
setting.

Probably I want to make it an abstract type. For PropBank and predicted
QA-SRL it makes more sense to use spans, but for Gold QA-SRL it makes
more sense to use questions with all of their answers, since multiple
spans were specifically given for each question (and not the other way
around), so it makes sense to treat the question as the instance.

We can also try just directly clustering the argument spans in gold
QA-SRL. But it's probably good to gradually migrate from the current
method to that way.