Add questions and make a small change to gapping
In order to make wh-questions about the subject go a bit more smoothly,
I decided to just make gaps into empty trees. This means you can't
locate the gap in a question; but also it helps make the question repr
ambiguous wrt the proposition repr, so could help in terms of
normalizing out the representation? idk